+++ packages/reactabular-easy/CHANGELOG.md
@@ -2,6 +2,13 @@
 ==================
 
   * Bug fix - Pass `parentField` to `tree.filter`. Now toggling should work.
+  * Breaking - Push `onMoveColumns` behavior out of table core. Now it doesn't maintain table state at all. As a result you need to implement `onMoveColumns` handler like this to make column moving work:
+
+```javascript
+onMoveColumns({ columns, source, target }) {
+  this.setState({ columns });
+}
+```
 
 6.1.2 / 2016-10-31
 ==================